extract helper functions to have quicker overview of actual tests
